Java Milestone 4 - Advanced Java/Milestone4FunctionalUnitTestExercises/src/main/java/gdb/milestone4functionalunittestexercises/TrimOne.java
package gdb.milestone4functionalunittestexercises;

/**
 * @date Thursday December 26, 2019
 * @author garrettbecker
 */

public class TrimOne {
    // Given a String, return a version without the first and 
    // last char, so "Hello" yields "ell". The String length will be at least 2. 
    //
    // trimOne("Hello") -> "ell"
    // trimOne("java") -> "av"
    // trimOne("coding") -> "odin"
    public String trimOne(String str) {
        int length = str.length();
        String toPrint = str.substring(1, length - 1);
        return toPrint;
    }
}
